+++ pkg/analysis_server/lib/src/services/correction/dart/abstract_producer.dart
@@ -249,7 +249,7 @@ class CorrectionProducerContext {
     this.overrideSet,
     this.selectionOffset = -1,
     this.selectionLength = 0,
-  })  : file = resolvedResult.path!,
+  })  : file = resolvedResult.path,
         session = resolvedResult.session,
         sessionHelper = AnalysisSessionHelper(resolvedResult.session),
         typeProvider = resolvedResult.typeProvider,