+++ runtime/tests/vm/dart_2/isolates/reload_many_isolates_live_and_die_test.dart
@@ -3,10 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:io';
-import 'dart:convert';
 
-import 'package:path/path.dart' as path;
 import 'package:expect/expect.dart';
 
 import 'reload_utils.dart';