+++ pkg/analyzer/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.0
 * Changed `AnalysisResult.path` to be non-nullable.
+* Changed `ParsedLibraryResult.units` to be non-nullable.
 
 ## 2.0.0
 * Removed deprecated `Scope.lookup2()`.